Split base-types.h and string-format.h out of common.h
This allows them to be used in headers such as `src/type.h`
which, because they are needed in common.h, cannot themselves
include `common.h`

This allows, for example, `StringPrintf` to be used in type.h
which is something I'm planning to do as a followup.
